pyproject.toml
[project]
name = "chart-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.12.0",
    "openai>=1.55.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "streamlit>=1.40.1",
    "tabulate>=0.9.0",
]
